Refactor project structure

+++ patients/list.php
@@ -12,25 +12,34 @@ $result = $conn->query("SELECT * FROM patients");
 
 <table border="1" cellpadding="10" cellspacing="0">
     <tr>
+        <th>ID</th>
         <th>Name</th>
-        <th>Age</th>
         <th>Gender</th>
+        <th>Age</th>
         <th>Phone</th>
+        <th>Address</th>
         <th>Action</th>
     </tr>
+<?php
 
-    <?php while($row = $result->fetch_assoc()) { ?>
+$sql = "SELECT * FROM patients";
+$result = $conn->query($sql);
+
+while ($row = $result->fetch_assoc()) {  ?> 
     <tr>
-        <td><?php echo $row['name']; ?></td>
-        <td><?php echo $row['age']; ?></td>
-        <td><?php echo $row['gender']; ?></td>
-        <td><?php echo $row['phone']; ?></td>
+        <td><?php   echo $row['id'] ?></td>
+        <td><?php   echo $row['full_name'] ?></td>
+        <td><?php   echo $row['gender'] ?></td>
+        <td><?php   echo $row['age'] ?></td>
+        <td><?php   echo $row['phone'] ?></td>
+        <td><?php   echo $row['address'] ?></td>
         <td>
-            <a href="edit.php?id=<?php echo $row['id']; ?>">Edit</a>
-            <a href="delete.php?id=<?php echo $row['id']; ?>">Delete</a>
+            <a href="edit.php?id=<?php echo $row['id']; ?>">Edit</a> |
+            <a href="delete.php?id=<?php echo $row['id']; ?>" onclick="return confirm('Are you sure?')">Delete</a>
         </td>
     </tr>
-    <?php } ?>
+    
+ <?php } ?>
 </table>
 
 <?php
